// src/types/navigation.d.ts
export type NavigationKegiatanType = {
  title: string,
  // href: string,
  description: string,
}

export type NavigationJurusanType = {
  href: string,
  jurusan: string,
  keahlihan: string
}

export type NavigationWebTechnologyType = {
  technology: string,
  href: string,
  icon: LucideIcon,
}

export type NavMenuHomeItemType = {
  href: string,
  title: string,
  description: string,
}

export type NavigationMenuMoreType = {
  href: string,
  name: string,
}